Adds KinematicBody
