+++ py_rpg/core/__init__.py
@@ -1,3 +1,4 @@
-__all__ = ["Node"]
+__all__ = ["Node", "KinematicBody"]
 
 from .node import Node
+from .kinematicbody import KinematicBody
